requirements.txt
flask
flask-session
requests
supabase
python-dotenv
gunicorn
